refactor: move cargo command to own file
